+++ src/functions.cpp
@@ -26,280 +26,208 @@ const char *LogToString(const Log::Level &str)
 	}
 }
 
-/**
- * @author DVT7125
- * @date 4/10/24
- * @brief Clears the screen of the controller and optionally the brain's screen.
- * @param brainClear Wither to clear the brain (`true`) or not (`false`).
- * @param controller1Clear Wither to clear controller1 (`true`) or not (`false`).
- */
-void clearScreen(const bool &brainClear, const bool &controller1Clear, const bool &controller2Clear)
+const char *LogToColor(const Log::Level &str)
 {
-	if (brainClear)
+	switch (str)
 	{
-		Brain.Screen.clearScreen();
-		Brain.Screen.setCursor(1, 1);
-	}
-	if (controller1Clear)
+	case Log::Level::Trace:
 	{
-		Controller1.Screen.clearScreen();
-		Controller1.Screen.setCursor(1, 1);
+		return "\033[92m[Trace]";
 	}
-	if (controller2Clear)
+
+	case Log::Level::Debug:
 	{
-		Controller2.Screen.clearScreen();
-		Controller2.Screen.setCursor(1, 1);
+		return "\033[93m[Debug]";
 	}
-	logHandler("clearScreen", "Finished clearScreen", Log::Level::Trace);
+
+	case Log::Level::Info:
+	{
+		return "\033[94m[Info]";
+	}
+
+	case Log::Level::Warn:
+	{
+		return " \033[38;5;216m[Warn]";
+	}
+
+	case Log::Level::Error:
+	{
+		return "\033[31m[Error]";
+	}
+
+	case Log::Level::Fatal:
+	{
+		return "\033[31m[Error]";
+	}
+	default:
+	{
+		return "\033[31m[Error]";
+	}
+	}
+}
+
+void calibrateGyro()
+{
+	InertialGyro.calibrate();
+	while (InertialGyro.isCalibrating())
+	{
+		vex::this_thread::sleep_for(20);
+	}
+	logHandler("calibrateGyro", "Finished calibrating Inertial Giro.", Log::Level::Trace);
 	return;
 }
 
-/**
- * @author @DVT7125
- * @date 4/10/24
- * @brief A module that handles controller button press.
- * @return A string representing the button pressed on the controller.
- */
-std::string ctrl1BttnPressed()
+std::pair<std::string, int> ctrl1BttnPressed()
 {
-	if (CONTROLLER1COMMAND)
+	if (Competition.isEnabled())
 	{
-		logHandler("ctrl1BttnPressed", "Ctrl1 is in command mode!", Log::Level::Error);
+		logHandler("ctrl1BttnPressed", "Robot is IN competition mode!", Log::Level::Error);
 	}
 
 	std::string buttonPressed;
-
-	while (!CONTROLLER1COMMAND)
+	int pressDuration = 0;
+	vex::timer pressTimer;
+	while (!Competition.isEnabled())
 	{
-		if (Controller1.ButtonA.pressing())
+		if (primaryController.ButtonA.pressing())
 		{
 			buttonPressed = "A";
+			pressTimer.clear();
+			while (primaryController.ButtonA.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonX.pressing())
+		else if (primaryController.ButtonX.pressing())
 		{
 			buttonPressed = "X";
+			pressTimer.clear();
+			while (primaryController.ButtonX.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonY.pressing())
+		else if (primaryController.ButtonY.pressing())
 		{
 			buttonPressed = "Y";
+			pressTimer.clear();
+			while (primaryController.ButtonY.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonB.pressing())
+		else if (primaryController.ButtonB.pressing())
 		{
 			buttonPressed = "B";
+			pressTimer.clear();
+			while (primaryController.ButtonB.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonUp.pressing())
+		else if (primaryController.ButtonUp.pressing())
 		{
 			buttonPressed = "UP";
+			pressTimer.clear();
+			while (primaryController.ButtonUp.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonDown.pressing())
+		else if (primaryController.ButtonDown.pressing())
 		{
 			buttonPressed = "DOWN";
+			pressTimer.clear();
+			while (primaryController.ButtonDown.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonRight.pressing())
+		else if (primaryController.ButtonRight.pressing())
 		{
 			buttonPressed = "RIGHT";
+			pressTimer.clear();
+			while (primaryController.ButtonRight.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonLeft.pressing())
+		else if (primaryController.ButtonLeft.pressing())
 		{
 			buttonPressed = "LEFT";
+			pressTimer.clear();
+			while (primaryController.ButtonLeft.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonL1.pressing())
+		else if (primaryController.ButtonL1.pressing())
 		{
 			buttonPressed = "L1";
+			pressTimer.clear();
+			while (primaryController.ButtonL1.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonL2.pressing())
+		else if (primaryController.ButtonL2.pressing())
 		{
 			buttonPressed = "L2";
+			pressTimer.clear();
+			while (primaryController.ButtonL2.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-
-		else if (Controller1.ButtonR1.pressing())
+		else if (primaryController.ButtonR1.pressing())
 		{
 			buttonPressed = "R1";
+			pressTimer.clear();
+			while (primaryController.ButtonR1.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
-		else if (Controller1.ButtonR2.pressing())
+		else if (primaryController.ButtonR2.pressing())
 		{
 			buttonPressed = "R2";
+			pressTimer.clear();
+			while (primaryController.ButtonR2.pressing())
+			{
+				vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
+			}
+			pressDuration = pressTimer.time();
 			break;
 		}
 
-		vex::this_thread::sleep_for(CTRLR1POLLINGRATE);
+		vex::this_thread::sleep_for(ConfigManager.getCtrlr1PollingRate());
 	}
 
-	std::string message = "Selected button: " + buttonPressed;
+	std::string message = "Selected button: " + buttonPressed + ", Duration: " + std::to_string(pressDuration) + " ms";
 	logHandler("ctrl1BttnPressed", message, Log::Level::Debug);
-	return buttonPressed;
-}
-
-/**
- * @author @DVT7125
- * @date 4/10/24
- * @brief Displays options and gets user input.
- * @param settingName Passed setting name.
- * @param options Passed options {"test", "test2"} (From 2 options to 4)
- * @return The user's choice.
- */
-std::string getUserOption(const std::string &settingName, const std::vector<std::string> &options)
-{
-	if (CONTROLLER1COMMAND)
-	{
-		logHandler("getUserOption", "Controller1 is IN command mode!", Log::Level::Error);
-	}
-
-	if (options.size() > MAXOPTIONSSIZE or options.size() < 2)
-	{
-		logHandler("getUserOption", "`options` size error!", Log::Level::Error);
-		return "DEFAULT";
-	}
-
-	std::ostringstream optmessage;
-	optmessage << "Options: ";
-	for (auto optionIterator = options.begin(); optionIterator != options.end(); ++optionIterator)
-	{
-		if (optionIterator != options.begin())
-		{
-			optmessage << ", "; // Separate options with a comma and space (except for the first option)
-		}
-		optmessage << *optionIterator;
-	}
-
-	logHandler("getUserOption", optmessage.str(), Log::Level::Debug);
-	optmessage.str(std::string());
-
-	std::size_t wrongAttemptCount = 0;
-	const std::size_t maxWrongAttempts = 3;
-	const std::string wrongMessages[maxWrongAttempts] = {
-		"Invalid selection!", "Do you need a map?", "Rocks can do this!"};
-	std::string buttonString;
-	// Have a index number for what they selected and convert it to the option.
-	std::size_t Index = options.size(); // Invalid selection by default
-	int offset = 0;
-
-	const std::vector<std::string> &buttons = {"A", "X", "Y", "B"};
-	const std::vector<std::string> &scrollButtons = {"DOWN", "UP"};
-
-	while (!Controller1.installed())
-	{
-		vex::this_thread::sleep_for(CTRLR1POLLINGRATE);
-	}
-
-	while (!CONTROLLER1COMMAND and Controller1.installed())
-	{
-		buttonString.clear();
-		clearScreen(false, true, true);
-		Controller1.Screen.print(settingName.c_str());
-
-		for (int i = 0; i < 2; ++i) // Checks for option size, and allows for options.
-		{
-			Controller1.Screen.newLine();
-			Controller1.Screen.print("%s: %s", buttons[i - offset].c_str(), options[i - offset].c_str());
-			buttonString += buttons[i - offset];
-			if (i != 1) // Add comma if not the last button
-			{
-				buttonString += ", ";
-			}
-		}
-
-		if ((options.size() == 3 and offset != -1) or (options.size() == 4 and offset != -2))
-		{
-			Controller1.Screen.setCursor(3, 24);
-			Controller1.Screen.print(">");
-		}
-		else if ((options.size() == 3 and offset == -1) or (options.size() == 4 and offset == -2))
-		{
-			Controller1.Screen.setCursor(3, 24);
-			Controller1.Screen.print("^");
-		}
-
-		optmessage << "Available buttons for current visible options: " << buttonString; // Append button string to message.
-		logHandler("getUserOption", optmessage.str(), Log::Level::Debug);
-		optmessage.str(std::string());
-
-		const std::string &buttonPressed = ctrl1BttnPressed(); // Get user input
-
-		if (buttonPressed == "A")
-		{
-			Index = 0;
-		}
-		else if (buttonPressed == "X" and options.size() >= 2)
-		{
-			Index = 1;
-		}
-		else if (buttonPressed == "Y" and options.size() >= 3)
-		{
-			Index = 2;
-		}
-		else if (buttonPressed == "B" and options.size() == 4)
-		{
-			Index = 3;
-		}
-
-		if (Index < options.size())
-		{
-			optmessage << "[Valid Selection] Index = " << Index << " | Offset = " << offset;
-			logHandler("get_User_Option", optmessage.str(), Log::Level::Debug);
-			break;
-		}
-
-		// Don't even ask what this does
-		else if (end(scrollButtons) != std::find(begin(scrollButtons), end(scrollButtons), buttonPressed))
-		{
-			if (buttonPressed == "DOWN" and ((options.size() == 3 and offset != -1) or (options.size() == 4 and offset != -2)))
-			{
-				--offset;
-			}
-			else if (buttonPressed == "UP" and offset != 0)
-			{
-				++offset;
-			}
-			optmessage << "[Valid Selection] Index = " << Index << " | Offset = " << offset;
-			logHandler("get_User_Option", optmessage.str(), Log::Level::Debug);
-		}
-
-		else
-		{
-			optmessage << "[Invalid Selection] Index = " << Index << " | Offset = " << offset;
-			logHandler("getUserOption", optmessage.str(), Log::Level::Debug);
-			// Display message
-			if (wrongAttemptCount < maxWrongAttempts)
-			{
-				clearScreen(false, true, true);
-				Controller1.Screen.print(wrongMessages[wrongAttemptCount].c_str());
-				++wrongAttemptCount; // Increment wrong attempt count
-				std::ostringstream failattemptdebug;
-				failattemptdebug << "wrongAttemptCount: " << wrongAttemptCount;
-				logHandler("getUserOption", failattemptdebug.str(), Log::Level::Debug);
-				vex::this_thread::sleep_for(2000);
-			}
-
-			else
-			{
-				logHandler("getUserOption", "Your half arsed.", Log::Level::Fatal);
-			}
-		}
-		optmessage.str(std::string());
-	}
-	clearScreen(false, true, true);
-	return options[Index];
+	return std::make_pair(buttonPressed, pressDuration);
 }
 
 /**
@@ -311,38 +239,42 @@ void motorTempMonitor()
 	std::ostringstream motorTemps;
 	std::ostringstream dataBuffer;
 
-	while (CONTROLLER1COMMAND)
+	while (Competition.isEnabled())
 	{
 		motorTemps.str(std::string());
 		dataBuffer.str(std::string());
 
-		int clawTemp = ClawMotor.temperature(vex::temperatureUnits::celsius);
-		int armTemp = ArmMotor.temperature(vex::temperatureUnits::celsius);
-		int leftDriveTemp = LeftDriveSmart.temperature(vex::temperatureUnits::celsius);
-		int rightDriveTemp = RightDriveSmart.temperature(vex::temperatureUnits::celsius);
+		int frontLeftTemp = frontLeftMotor.temperature(vex::temperatureUnits::celsius);
+		int frontRightTemp = frontRightMotor.temperature(vex::temperatureUnits::celsius);
+		int rearLeftTemp = rearLeftMotor.temperature(vex::temperatureUnits::celsius);
+		int rearRightTemp = rearRightMotor.temperature(vex::temperatureUnits::celsius);
+
+		// Get the position of both motors
+		int leftMotorPosition = LeftDriveSmart.position(vex::rotationUnits::deg);
+		int rightMotorPosition = RightDriveSmart.position(vex::rotationUnits::deg);
 
 		// Check for overheat conditions for each motor
-		if (clawTemp >= 55)
+		if (frontLeftTemp >= 55)
 		{
-			motorTemps << "CM overheat: " << clawTemp << "°";
+			motorTemps << "FLM overheat: " << frontLeftTemp << "°";
 			logHandler("motorTempMonitor", motorTemps.str(), Log::Level::Warn);
 			motorTemps.str(std::string());
 		}
-		if (armTemp >= 55)
+		if (frontRightTemp >= 55)
 		{
-			motorTemps << "AM overheat: " << armTemp << "°";
+			motorTemps << "FRM overheat: " << frontRightTemp << "°";
 			logHandler("motorTempMonitor", motorTemps.str(), Log::Level::Warn);
 			motorTemps.str(std::string());
 		}
-		if (leftDriveTemp >= 55)
+		if (rearLeftTemp >= 55)
 		{
-			motorTemps << "LM overheat: " << leftDriveTemp << "°";
+			motorTemps << "RLM overheat: " << rearLeftTemp << "°";
 			logHandler("motorTempMonitor", motorTemps.str(), Log::Level::Warn);
 			motorTemps.str(std::string());
 		}
-		if (rightDriveTemp >= 55)
+		if (rearRightTemp >= 55)
 		{
-			motorTemps << "RM overheat: " << rightDriveTemp << "°";
+			motorTemps << "RRM overheat: " << rearRightTemp << "°";
 			logHandler("motorTempMonitor", motorTemps.str(), Log::Level::Warn);
 			motorTemps.str(std::string());
 		}
@@ -350,51 +282,52 @@ void motorTempMonitor()
 		{
 			logHandler("motorTempMonitor", "Brain voltage at a critical level! Performance will be reduced!", Log::Level::Warn);
 		}
+		// Get the average of the two motors
+		int averagePosition = (leftMotorPosition + rightMotorPosition) / 2;
+
+		// Update the odometer with the new position
+		ConfigManager.updateOdometer(averagePosition);
 
 		// Log motor temperatures
-		motorTemps << "\n | LeftTemp: " << leftDriveTemp << "°\n | RightTemp: " << rightDriveTemp << "°\n | ArmTemp: " << armTemp << "°\n | ClawTemp: " << clawTemp << "°\n | Battery Voltage: " << Brain.Battery.voltage() << "V\n";
+		motorTemps << "\n | LeftTemp: " << frontLeftTemp << "°\n | RightTemp: " << frontRightTemp << "°\n | RearLeftTemp: " << rearLeftTemp << "°\n | RearRightTemp: " << rearRightTemp << "°\n | Battery Voltage: " << Brain.Battery.voltage() << "V\n";
 		logHandler("motorTempMonitor", motorTemps.str(), Log::Level::Info);
-		dataBuffer << "\nX Axis: " << Inertial.pitch(vex::rotationUnits::deg) << "\nY Axis: " << Inertial.roll(vex::rotationUnits::deg) << "\nZ Axis: " << Inertial.yaw(vex::rotationUnits::deg);
+		dataBuffer << "\nX Axis: " << InertialGyro.pitch(vex::rotationUnits::deg) << "\nY Axis: " << InertialGyro.roll(vex::rotationUnits::deg) << "\nZ Axis: " << InertialGyro.yaw(vex::rotationUnits::deg);
 		logHandler("motorTempMonitor", dataBuffer.str(), Log::Level::Info);
 		clearScreen(false, true, true);
-		Controller1.Screen.print("LM: %d° | RM: %d°", leftDriveTemp, rightDriveTemp);
-		Controller1.Screen.newLine();
-		Controller1.Screen.print("AM: %d° | CM: %d°", armTemp, clawTemp);
-		Controller1.Screen.newLine();
-		Controller1.Screen.print("Battery: %.1fV", Brain.Battery.voltage());
+		primaryController.Screen.print("FLM: %d° | FRM: %d°", frontLeftTemp, frontRightTemp);
+		primaryController.Screen.newLine();
+		primaryController.Screen.print("RLM: %d° | RRM: %d°", rearLeftTemp, rearRightTemp);
+		primaryController.Screen.newLine();
+		primaryController.Screen.print("Battery: %.1fV", Brain.Battery.voltage());
 		vex::this_thread::sleep_for(5000);
 	}
 	return;
 }
 
-/**
- * @brief Display GIF animations based on the value of drivercontrollogo.
- */
 void gifplayer()
 {
-	vex::Gif gif("assets/warp.gif", 0, 0);
-	
-	while (LOCALLOGO)
+	if (!Competition.isEnabled())
 	{
-		Brain.Screen.printAt(5, 200, "frame %3d", gif.getFrameIndex());
+		vex::Gif gif("assets/loading.gif", 0, 0);
+		while (!Competition.isEnabled())
+		{
+			Brain.Screen.print("");
+		}
 	}
-	return;
-}
-
-/**
- * @brief Calibrate the Inertial Gyro.
- *
- * @return int Always returns 1.
- */
-void calibrategiro()
-{
-	logHandler("calibrateDrivetrain", "Calibrating Inertial Gyro...", Log::Level::Info);
-	Inertial.calibrate();
-
-	while (Inertial.isCalibrating())
+	else if (Competition.isAutonomous())
 	{
-		vex::this_thread::sleep_for(25);
+		vex::Gif gif("assets/auto.gif", 0, 0);
+		while (Competition.isAutonomous())
+		{
+			Brain.Screen.print("");
+		}
 	}
-	logHandler("calibrateDrivetrain", "Finished Calibrating Inertial Gyro.", Log::Level::Info);
-	return;
+	else
+	{
+		vex::Gif gif("assets/driver.gif", 0, 0);
+		while (Competition.isDriverControl())
+		{
+			Brain.Screen.print("");
+		}
+	}
 }